moved precon utils
